pyproject.toml
[tool.poetry]
name = "flask-code"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Jvras <jonathas.vinicius@ufpe.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-restful = "^0.3.10"
flask-cors = "^5.0.0"
dynaconf = "^3.2.6"
rich = "^13.9.4"
flask-restx = "^1.3.0"
flask-migrate = "^4.0.7"
flask-sqlalchemy = "^3.1.1"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
mock = "^5.1.0"



[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.blue]
extend-exclude = '(migrations/)'
line_length = 88

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
